backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
openai==1.93.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.0
beautifulsoup4==4.12.2
requests==2.32.4
newspaper3k==0.2.8
trafilatura==2.0.0
lxml_html_clean==0.4.2